requirements.txt
pyserial>=3.5
flask
rethinkdb
pyudev==0.24.1